requirements.txt
astroid==2.11.4
attrs==21.4.0
bandit==1.7.4
black==22.3.0
captcha==0.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==6.3.2
dill==0.3.4
Django==2.2
dj-database-url==0.5.0
django-braces==1.15.0
django-environ==0.8.1
django-model-utils==4.2.0
django-phonenumber-field==6.1.0
django-recaptcha==3.0.0
django-star-ratings==0.9.2
djongo==1.3.4
dnspython==2.2.1
dodgy==0.2.1
eradicate==2.1.0
flake8==4.0.1
flake8-eradicate==1.2.1
flake8-polyfill==1.0.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
imagekitio==2.2.8
isort==5.10.1
lazy-object-proxy==1.7.1
mando==0.6.4
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
pbr==5.9.0
pep8-naming==0.10.0
phonenumbers==8.12.47
Pillow==9.1.0
platformdirs==2.5.2
prospector==1.7.7
psycopg2-binary==2.9.3
pybase64==1.2.1
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
pylint==2.13.8
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pymongo==3.12.3
python-dateutil==1.5
pytz==2021.3
PyYAML==6.0
radon==5.1.0
recaptcha==1.0rc1
requests==2.27.1
requirements-detector==0.7
setoptconf-tmp==0.3.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sqlparse==0.2.4
stevedore==3.5.0
swapper==1.3.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.9
whitenoise==6.1.0
wrapt==1.14.1
